# #1459 - Faster support vector/function for BallInf