--- a/docs/src/lib/representations.md
+++ b/docs/src/lib/representations.md
@@ -46,6 +46,7 @@ radius_hyperrectangle(::BallInf{N}, ::Int) where {N<:Real}
 isflat(::BallInf)
 rand(::Type{BallInf})
 σ(::AbstractVector{N}, ::BallInf{N}) where {N<:Real}
+ρ(::AbstractVector{N}, ::BallInf{N}) where {N<:Real}
 translate(::BallInf{N}, ::AbstractVector{N}) where {N<:Real}
 ```
 Inherited from [`LazySet`](@ref):
@@ -64,7 +65,6 @@ Inherited from [`AbstractZonotope`](@ref):
 * [`ngens`](@ref ngens(::AbstractZonotope))
 
 Inherited from [`AbstractHyperrectangle`](@ref):
-* [`ρ`](@ref ρ(::AbstractVector{N}, ::AbstractHyperrectangle{N}) where {N<:Real})
 * [`∈`](@ref ∈(::AbstractVector{N}, ::AbstractHyperrectangle{N}) where {N<:Real})
 * [`norm`](@ref norm(::AbstractHyperrectangle, ::Real))
 * [`vertices_list`](@ref vertices_list(::AbstractHyperrectangle{N}) where {N<:Real})