--- a/src/BallInf.jl
+++ b/src/BallInf.jl
@@ -164,6 +164,62 @@ function σ(d::AbstractVector{N}, B::BallInf{N}) where {N<:Real}
 end
 
 """
+    ρ(d::AbstractVector{N}, B::BallInf{N}) where {N<:Real}
+
+Evaluate the support function of a ball in the infinity norm in a given
+direction.
+
+### Input
+
+- `d` -- direction
+- `B` -- ball in the infinity norm
+
+### Output
+
+Evaluation of the support function in the given direction.
+
+### Algorithm
+
+Let ``B`` be a ball in the infinity norm with center ``c`` and radius ``r`` and
+let `d` be the direction of interest.
+For balls with dimensions less than 30 we use the implementation for
+`AbstractHyperrectangle`, taylored to a `BallInf`, which computes
+
+```math
+    ∑_{i=1}^n d_i * (c_i + \\sgn(d_i) * r)
+```
+
+where ``\\sgn(α) = 1`` if ``α ≥ 0`` and ``\\sgn(α) = 1`` if ``α < 0``.
+
+For balls of higher dimension, we instead exploit that for a support vector
+``v = σ(d, B) = c + \\sgn(d) * (r, …, r)ᵀ`` we have
+
+```math
+    ρ(d, B) = ⟨d, v⟩ = ⟨d, c⟩ + ⟨d, \\sgn(d) * (r, …, r)ᵀ⟩ = ⟨d, c⟩ + r · ∑_{i=1}^n |d_i|
+```
+
+where ``⟨·, ·⟩`` denotes the dot product.
+"""
+function ρ(d::AbstractVector{N}, B::BallInf{N}) where {N<:Real}
+    @assert length(d) == dim(B) "a $(length(d))-dimensional vector is " *
+                                "incompatible with a $(dim(B))-dimensional set"
+    c = center(B)
+    if length(d) > 30
+        # more efficient for higher dimensions
+        return dot(d, c) + B.radius * sum(abs, d)
+    end
+    res = zero(N)
+    @inbounds for (i, di) in enumerate(d)
+        if di < zero(N)
+            res += di * (c[i] - B.radius)
+        elseif di > zero(N)
+            res += di * (c[i] + B.radius)
+        end
+    end
+    return res
+end
+
+"""
     radius(B::BallInf, [p]::Real=Inf)::Real
 
 Return the radius of a ball in the infinity norm.