Closes #1459.

```julia
julia> using LazySets, BenchmarkTools
julia> for n in [1, 10, 100, 1000]
           B = rand(BallInf, dim=n)
           @btime σ($(ones(n)), $B)
       end
  # master
  79.740 ns (2 allocations: 192 bytes)
  97.977 ns (2 allocations: 320 bytes)
  236.966 ns (2 allocations: 1.75 KiB)
  2.022 μs (2 allocations: 15.88 KiB)
  # this branch
  48.857 ns (1 allocation: 96 bytes)
  61.066 ns (1 allocation: 160 bytes)
  116.163 ns (1 allocation: 896 bytes)
  971.600 ns (1 allocation: 7.94 KiB)
```