@schillic 2019-06-22
While thinking about the support function of the `Ball2`, I found a faster way to compute the support function of `BallInf`, which I added here as a bonus.

```julia
julia> using LazySets, BenchmarkTools
julia> for n in [1, 10, 100, 1000]
           B = rand(BallInf, dim=n)
           @btime ρ($(ones(n)), $B)
       end
  # master
  7.089 ns (0 allocations: 0 bytes)
  17.147 ns (0 allocations: 0 bytes)
  156.205 ns (0 allocations: 0 bytes)
  1.497 μs (0 allocations: 0 bytes)
  # this branch
  19.178 ns (0 allocations: 0 bytes)
  28.566 ns (0 allocations: 0 bytes)
  62.421 ns (0 allocations: 0 bytes)
  302.040 ns (0 allocations: 0 bytes)
```